🎨 Merge sign in methods to header
