+++ components/WalletHeader.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="flex flex-wrap items-center gap-2">
-    <template v-if="wallet">
+    <template v-if="bookStoreApiStore.isAuthenticated">
       <UTooltip :text="wallet">
         <UButton
           class="text-xs font-mono"
@@ -23,7 +23,7 @@
       label="Sign in"
       icon="i-heroicons-arrow-right-on-rectangle"
       color="primary"
-      @click="connect"
+      @click="onClickAuth"
     />
   </div>
 </template>
@@ -33,12 +33,20 @@ import { storeToRefs } from 'pinia'
 import { useWalletStore } from '~/stores/wallet'
 import { getPortfolioURL } from '~/utils'
 import { shortenWalletAddress } from '~/utils/cosmos'
+import { useBookStoreApiStore } from '~/stores/book-store-api'
 
 const store = useWalletStore()
-const { wallet } = storeToRefs(store)
-const { connect, disconnect } = store
+const { wallet, signer } = storeToRefs(store)
+const { connect, disconnect, signMessageMemo } = store
+const bookStoreApiStore = useBookStoreApiStore()
+const { authenticate } = bookStoreApiStore
+const { token, wallet: sessionWallet } = storeToRefs(bookStoreApiStore)
+
+const toast = useToast()
 
 const portfolioURL = computed(() => getPortfolioURL(wallet.value))
+
+const isLoading = ref(false)
 
 onMounted(async () => {
   try {
@@ -49,6 +57,39 @@ onMounted(async () => {
     }
   } catch {}
 })
+
+async function onClickAuth () {
+  try {
+    isLoading.value = true
+    if (!wallet.value || !signer.value) {
+      await connect()
+    }
+    if (!wallet.value || !signer.value) { return }
+    const signature = await signMessageMemo(
+      'authorize',
+      ['read:nftbook', 'write:nftbook', 'read:nftcollection', 'write:nftcollection']
+    )
+    if (!signature) { return }
+    await authenticate(wallet.value, signature)
+    try {
+      window.localStorage.setItem('likecoin_nft_book_press_token', JSON.stringify({ wallet: sessionWallet.value, token: token.value }))
+    } catch (err) {}
+  } catch (err) {
+    // eslint-disable-next-line no-console
+    console.error(err)
+    toast.add({
+      icon: 'i-heroicons-exclamation-circle',
+      title: (err as Error).toString(),
+      timeout: 0,
+      color: 'red',
+      ui: {
+        title: 'text-red-400 dark:text-red-400'
+      }
+    })
+  } finally {
+    isLoading.value = false
+  }
+}
 
 function onClickDisconnect () {
   disconnect()
